+++ src/beamngpy/sensors/GPS.py
@@ -1,12 +1,11 @@
 from __future__ import annotations
 
 from logging import DEBUG, getLogger
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
-from beamngpy.logging import LOGGER_ID, BNGError
+from beamngpy.connection import CommBase
+from beamngpy.logging import LOGGER_ID
 from beamngpy.types import Float3, StrDict
-
-from .communication_utils import send_sensor_request, set_sensor
 
 if TYPE_CHECKING:
     from beamngpy.beamng import BeamNGpy
@@ -15,7 +14,7 @@ if TYPE_CHECKING:
 __all__ = ['GPS']
 
 
-class GPS:
+class GPS(CommBase):
     """
     This automated sensor provides GPS readings (position) in spherical coordinates (lattitude, longitude).  It can be attached to any point on or relative to the vehicle.
 
@@ -34,36 +33,26 @@ class GPS:
     """
 
     def __init__(self, name: str, bng: BeamNGpy, vehicle: Vehicle, gfx_update_time: float = 0.0, physics_update_time: float = 0.01, pos: Float3 = (0, 0, 1.7),
-        ref_lon: float = 0.0, ref_lat: float = 0.0, is_send_immediately: bool = False, is_visualised: bool = True, is_snapping_desired: bool = False,
-        is_force_inside_triangle: bool = False):
+                 ref_lon: float = 0.0, ref_lat: float = 0.0, is_send_immediately: bool = False, is_visualised: bool = True, is_snapping_desired: bool = False,
+                 is_force_inside_triangle: bool = False):
+        super().__init__(bng, vehicle)
 
         self.logger = getLogger(f'{LOGGER_ID}.GPS')
         self.logger.setLevel(DEBUG)
 
         # Cache some properties we will need later.
-        self.bng = bng
         self.name = name
-        self.vehicle = vehicle
         self.is_send_immediately = is_send_immediately
+        self.vehicle = vehicle
 
         # Create and initialise this sensor in the simulation.
         self._open_GPS(name, vehicle, gfx_update_time, physics_update_time, pos, ref_lon, ref_lat, is_send_immediately,
-            is_visualised, is_snapping_desired, is_force_inside_triangle)
+                       is_visualised, is_snapping_desired, is_force_inside_triangle)
 
         # Fetch the unique Id number (in the simulator) for this GPS sensor.  We will need this later.
         self.sensorId = self._get_GPS_id()
 
         self.logger.debug('GPS - sensor created: 'f'{self.name}')
-
-    def _send_sensor_request(self, type: str, ack: str | None = None, **kwargs: Any) -> StrDict:
-        if not self.bng.connection:
-            raise BNGError('The simulator is not connected!')
-        return send_sensor_request(self.bng.connection, type, ack, **kwargs)
-
-    def _set_sensor(self, type: str, ack: str | None = None, **kwargs: Any) -> None:
-        if not self.bng.connection:
-            raise BNGError('The simulator is not connected!')
-        set_sensor(self.bng.connection, type, **kwargs)
 
     def remove(self) -> None:
         """
@@ -103,7 +92,7 @@ class GPS:
             A unique Id number for the ad-hoc request.
         """
         self.logger.debug('GPS - ad-hoc polling request sent: 'f'{self.name}')
-        return int(self._send_sensor_request('SendAdHocRequestGPS', 'CompletedSendAdHocRequestGPS', name=self.name, vid=self.vehicle.vid)['data'])
+        return int(self.send_recv_ge('SendAdHocRequestGPS', name=self.name, vid=self.vehicle.vid)['data'])
 
     def is_ad_hoc_poll_request_ready(self, request_id: int) -> bool:
         """
@@ -116,7 +105,7 @@ class GPS:
             A flag which indicates if the ad-hoc polling request is complete.
         """
         self.logger.debug('GPS - ad-hoc polling request checked for completion: 'f'{self.name}')
-        return self._send_sensor_request('IsAdHocPollRequestReadyGPS', 'CompletedIsAdHocPollRequestReadyGPS', requestId=request_id)['data']
+        return self.send_recv_ge('IsAdHocPollRequestReadyGPS', requestId=request_id)['data']
 
     def collect_ad_hoc_poll_request(self, request_id: int) -> StrDict:
         """
@@ -128,7 +117,7 @@ class GPS:
         Returns:
             The readings data.
         """
-        readings = self._send_sensor_request('CollectAdHocPollRequestGPS', 'CompletedCollectAdHocPollRequestGPS', requestId=request_id)['data']
+        readings = self.send_recv_ge('CollectAdHocPollRequestGPS', requestId=request_id)['data']
         self.logger.debug('GPS - ad-hoc polling request returned and processed: 'f'{self.name}')
         return readings
 
@@ -139,7 +128,8 @@ class GPS:
         Args:
             requested_update_time: The new requested update time.
         """
-        self._set_sensor('SetGPSRequestedUpdateTime', ack='CompletedSetGPSRequestedUpdateTime', name=self.name, vid=self.vehicle.vid, GFXUpdateTime=requested_update_time)
+        self.send_ack_ge('SetGPSRequestedUpdateTime', ack='CompletedSetGPSRequestedUpdateTime',
+                         name=self.name, vid=self.vehicle.vid, GFXUpdateTime=requested_update_time)
 
     def set_is_visualised(self, is_visualised: bool) -> None:
         """
@@ -148,15 +138,15 @@ class GPS:
         Args:
             is_visualised: A flag which indicates if this sensor is to be visualised or not.
         """
-        self._set_sensor('SetGPSIsVisualised', ack='CompletedSetGPSIsVisualised', name=self.name, vid=self.vehicle.vid, isVisualised=is_visualised)
+        self.send_ack_ge('SetGPSIsVisualised', ack='CompletedSetGPSIsVisualised',
+                         name=self.name, vid=self.vehicle.vid, isVisualised=is_visualised)
 
     def _get_GPS_id(self) -> int:
-        return int(
-            self._send_sensor_request('GetGPSId', ack='CompletedGetGPSId', name=self.name)['data'])
+        return int(self.send_recv_ge('GetGPSId', name=self.name)['data'])
 
     def _open_GPS(self, name: str, vehicle: Vehicle, gfx_update_time: float, physics_update_time: float, pos: Float3, ref_lon: float, ref_lat: float,
-        is_send_immediately: bool, is_visualised: bool, is_snapping_desired: bool, is_force_inside_triangle: bool) -> None:
-        data: StrDict = dict(type='OpenGPS')
+                  is_send_immediately: bool, is_visualised: bool, is_snapping_desired: bool, is_force_inside_triangle: bool) -> None:
+        data: StrDict = dict()
         data['name'] = name
         data['vid'] = vehicle.vid
         data['GFXUpdateTime'] = gfx_update_time
@@ -168,20 +158,15 @@ class GPS:
         data['isVisualised'] = is_visualised
         data['isSnappingDesired'] = is_snapping_desired
         data['isForceInsideTriangle'] = is_force_inside_triangle
-        self.bng._send(data).ack('OpenedGPS')
+        self.send_ack_ge(type='OpenGPS', ack='OpenedGPS', **data)
         self.logger.info(f'Opened GPS sensor: "{name}"')
 
     def _close_GPS(self) -> None:
-        data = dict(type='CloseGPS')
-        data['name'] = self.name
-        data['vid'] = self.vehicle.vid
-        self.bng._send(data).ack('ClosedGPS')
+        self.send_ack_ge(type='CloseGPS', ack='ClosedGPS', name=self.name, vid=self.vehicle.vid)
         self.logger.info(f'Closed GPS sensor: "{self.name}"')
 
     def _poll_GPS_GE(self) -> StrDict:
-        return self._send_sensor_request('PollGPSGE', ack='PolledGPSGECompleted', name=self.name)['data']
+        return self.send_recv_ge('PollGPSGE', name=self.name)['data']
 
     def _poll_GPS_VE(self) -> StrDict:
-        if not self.vehicle.connection:
-            raise BNGError('The vehicle is not connected!')
-        return send_sensor_request(self.vehicle.connection, 'PollGPSVE', name=self.name, sensorId=self.sensorId)['data']
+        return self.send_recv_veh('PollGPSVE', name=self.name, sensorId=self.sensorId)['data']